fix(activity): derive unread badges from the Agents page thread build
The sidebar Agents pill and the Agents view computed "unread" from
different code paths, so the pill could keep showing a count that
"Mark all read" never cleared: manually-flagged workspace cards,
panes without tab context, and retained/event-cap asymmetries were
counted by the badge but never listed (or ackable) in the view.

- extract the event/thread builders out of ActivityPrototypePage into
  activity-events.ts + agent-pane-threads.ts (the page is lazy-loaded,
  so the always-mounted sidebar could not import from it)
- count unread threads via countUnreadAgentPaneThreads in
  useActivityUnreadCount for both the sidebar pill and the Agents
  titlebar badge; workspace-card isUnread no longer feeds the pill
- persist the unread-only toggle (orca.activity.read-filter) so the
  Agents view restores the filter across opens
