+++ src/renderer/src/components/activity/ActivityPrototypePage.test.ts
@@ -17,12 +17,11 @@ import {
   isActivityFilterFocusShortcut,
   shouldIgnoreActivityFilterFocusShortcutTarget,
   buildActivityThreadGroups,
-  buildActivityEvents,
-  buildAgentPaneThreads,
   getActivityThreadGroup,
   groupActivityThreadsByStatus,
   isActivitySearchQueryTooLarge
 } from './ActivityPrototypePage'
+import { buildActivityEvents, buildAgentPaneThreads } from './agent-pane-threads'
 import { makePaneKey } from '../../../../shared/stable-pane-id'
 
 const LEAF_ID = '11111111-1111-4111-8111-111111111111'
