+++ src/renderer/src/components/activity/useActivityUnreadCount.ts
@@ -1,107 +1,37 @@
 import { useMemo } from 'react'
 import { useShallow } from 'zustand/react/shallow'
 
-import { migrationUnsupportedToAgentStatusEntry } from '@/lib/migration-unsupported-agent-entry'
 import { useAppStore } from '@/store'
+import { getRepoMapFromState, getWorktreeMapFromState } from '@/store/selectors'
 import type { AppState } from '@/store/types'
-import type { AgentStatusEntry, AgentStatusState } from '../../../../shared/agent-status-types'
+import { countUnreadAgentPaneThreads } from './agent-pane-threads'
 
-type ActivityUnreadCountSource = Pick<
-  AppState,
-  | 'acknowledgedAgentsByPaneKey'
-  | 'agentStatusByPaneKey'
-  | 'migrationUnsupportedByPtyId'
-  | 'retainedAgentsByPaneKey'
-  | 'worktreesByRepo'
->
-
-type ActivityUnreadCountMode = 'agent-events' | 'sidebar-badge'
-
-const EMPTY_WORKTREES_BY_REPO: AppState['worktreesByRepo'] = {}
+const EMPTY_TABS_BY_WORKTREE: AppState['tabsByWorktree'] = {}
+const EMPTY_WORKTREE_MAP: ReturnType<typeof getWorktreeMapFromState> = new Map()
+const EMPTY_REPO_MAP: ReturnType<typeof getRepoMapFromState> = new Map()
 const EMPTY_MIGRATION_UNSUPPORTED: AppState['migrationUnsupportedByPtyId'] = {}
 const EMPTY_RETAINED_AGENTS: AppState['retainedAgentsByPaneKey'] = {}
 const EMPTY_ACKNOWLEDGED_AGENTS: AppState['acknowledgedAgentsByPaneKey'] = {}
 
 const DISABLED_ACTIVITY_UNREAD_INPUTS = {
   sortEpoch: 0,
-  worktreesByRepo: EMPTY_WORKTREES_BY_REPO,
+  tabsByWorktree: EMPTY_TABS_BY_WORKTREE,
+  worktreeMap: EMPTY_WORKTREE_MAP,
+  repoMap: EMPTY_REPO_MAP,
   migrationUnsupportedByPtyId: EMPTY_MIGRATION_UNSUPPORTED,
   retainedAgentsByPaneKey: EMPTY_RETAINED_AGENTS,
   acknowledgedAgentsByPaneKey: EMPTY_ACKNOWLEDGED_AGENTS
 }
 
-function isUnreadAgentState(state: AgentStatusState): boolean {
-  return state === 'done' || state === 'blocked' || state === 'waiting'
-}
-
-export function countActivityUnread(
-  source: ActivityUnreadCountSource,
-  mode: ActivityUnreadCountMode
-): number {
-  let count = 0
-
-  if (mode === 'sidebar-badge') {
-    for (const worktrees of Object.values(source.worktreesByRepo)) {
-      for (const worktree of worktrees) {
-        if (worktree.createdAt && worktree.isUnread) {
-          count += 1
-        }
-      }
-    }
-  }
-
-  const countEntry = (entry: AgentStatusEntry, ackAt: number): void => {
-    if (mode === 'agent-events') {
-      // Why: Activity feed surfaces historical done/blocked/waiting events
-      // from stateHistory, so the titlebar badge must mirror that event count.
-      for (const history of entry.stateHistory) {
-        if (isUnreadAgentState(history.state) && ackAt < history.startedAt) {
-          count += 1
-        }
-      }
-    }
-    // Why: a session-boundary done is an idle connect (STA-3386), not an event to read.
-    // History never contains a boundary, but it DOES keep the real completion a boundary
-    // displaced (the slice pushes it on done→done), so sidebar-badge mode — which skips the
-    // history loop above — must still count that displaced completion or the badge silently
-    // drops an unacknowledged finish the moment its session is resumed.
-    if (
-      isUnreadAgentState(entry.state) &&
-      entry.sessionBoundary !== true &&
-      ackAt < entry.stateStartedAt
-    ) {
-      count += 1
-    } else if (mode === 'sidebar-badge' && entry.state === 'done' && entry.sessionBoundary) {
-      const displaced = entry.stateHistory.at(-1)
-      if (displaced && isUnreadAgentState(displaced.state) && ackAt < displaced.startedAt) {
-        count += 1
-      }
-    }
-  }
-
-  for (const [paneKey, entry] of Object.entries(source.agentStatusByPaneKey)) {
-    countEntry(entry, source.acknowledgedAgentsByPaneKey[paneKey] ?? 0)
-  }
-  for (const [paneKey, retained] of Object.entries(source.retainedAgentsByPaneKey)) {
-    if (mode === 'sidebar-badge' && retained.entry.state !== 'done') {
-      continue
-    }
-    countEntry(retained.entry, source.acknowledgedAgentsByPaneKey[paneKey] ?? 0)
-  }
-  for (const unsupported of Object.values(source.migrationUnsupportedByPtyId)) {
-    const entry = migrationUnsupportedToAgentStatusEntry(unsupported)
-    if (entry) {
-      countEntry(entry, source.acknowledgedAgentsByPaneKey[entry.paneKey] ?? 0)
-    }
-  }
-
-  return count
-}
-
-export function useActivityUnreadCount(enabled: boolean, mode: ActivityUnreadCountMode): number {
+// Why: the sidebar pill and Agents titlebar badge must equal the unread rows the
+// Agents page lists (and that "Mark all read" acknowledges), so the count derives
+// from the same thread build instead of a parallel per-entry tally.
+export function useActivityUnreadCount(enabled: boolean): number {
   const {
     sortEpoch,
-    worktreesByRepo,
+    tabsByWorktree,
+    worktreeMap,
+    repoMap,
     migrationUnsupportedByPtyId,
     retainedAgentsByPaneKey,
     acknowledgedAgentsByPaneKey
@@ -115,7 +45,9 @@ export function useActivityUnreadCount(enabled: boolean, mode: ActivityUnreadCou
         // cannot change unread count unless a sort-relevant state transition
         // or removal occurred. sortEpoch is the cheap invalidation signal.
         sortEpoch: state.sortEpoch,
-        worktreesByRepo: state.worktreesByRepo,
+        tabsByWorktree: state.tabsByWorktree,
+        worktreeMap: getWorktreeMapFromState(state),
+        repoMap: getRepoMapFromState(state),
         migrationUnsupportedByPtyId: state.migrationUnsupportedByPtyId,
         retainedAgentsByPaneKey: state.retainedAgentsByPaneKey,
         acknowledgedAgentsByPaneKey: state.acknowledgedAgentsByPaneKey
@@ -128,23 +60,24 @@ export function useActivityUnreadCount(enabled: boolean, mode: ActivityUnreadCou
       return 0
     }
     void sortEpoch
-    return countActivityUnread(
-      {
-        agentStatusByPaneKey: useAppStore.getState().agentStatusByPaneKey,
-        migrationUnsupportedByPtyId,
-        retainedAgentsByPaneKey,
-        worktreesByRepo,
-        acknowledgedAgentsByPaneKey
-      },
-      mode
-    )
+    return countUnreadAgentPaneThreads({
+      agentStatusByPaneKey: useAppStore.getState().agentStatusByPaneKey,
+      migrationUnsupportedByPtyId,
+      retainedAgentsByPaneKey,
+      tabsByWorktree,
+      worktreeMap,
+      repoMap,
+      acknowledgedAgentsByPaneKey,
+      now: Date.now()
+    })
   }, [
     acknowledgedAgentsByPaneKey,
     enabled,
     migrationUnsupportedByPtyId,
-    mode,
+    repoMap,
     retainedAgentsByPaneKey,
     sortEpoch,
-    worktreesByRepo
+    tabsByWorktree,
+    worktreeMap
   ])
 }
